Updated link.sh to skip already linked files

+++ link.sh
@@ -2,37 +2,38 @@
 
 dotfiles=$HOME/.dotfiles/files/
 
-if [ -f ~/.bashrc ]; then
+if [ -f $HOME/.bashrc ] && [ ! -L $HOME/.bashrc ] ; then
     mv $HOME/.bashrc $HOME/.bashrc~
+    ln -s $dotfiles/bashrc $HOME/.bashrc
 fi
-ln -s $dotfiles/bashrc $HOME/.bashrc
 
-if [ -f $HOME/.forward ]; then
+if [ -f $HOME/.forward ] && [ ! -L $HOME/.foward ]; then
     mv $HOME/.forward $HOME/.forward~
+    ln -s $dotfiles/forward $HOME/.forward
 fi
-ln -s $dotfiles/forward $HOME/.forward
 
-if [ -f $HOME/.git ]; then
+if [ -f $HOME/.git ] && [ ! -L $HOME/.git ]; then
     mv $HOME/.git $HOME/.git~
+    ln -s $dotfiles/git $HOME/.git
 fi
-ln -s $dotfiles/git $HOME/.git
 
-if [ -f $HOME/.gitconfig ]; then
+if [ -f $HOME/.gitconfig ] && [ ! -L $HOME/.gitconfig ]; then
     mv $HOME/.gitconfig $HOME/.gitconfig~
+    ln -s $dotfiles/gitconfig $HOME/.gitconfig
 fi
-ln -s $dotfiles/gitconfig $HOME/.gitconfig
 
-if [ -f $HOME/.gitignore ]; then
+if [ -f $HOME/.gitignore ] && [ ! -L $HOME/.gitignore ]; then
     mv $HOME/.gitignore $HOME/.gitignore~
+    ln -s $dotfiles/gitignore $HOME/.gitignore
 fi
-ln -s $dotfiles/gitignore $HOME/.gitignore
 
-if [ -f $HOME/.selected_editor ]; then
+if [ -f $HOME/.selected_editor ] && [ ! -L $HOME/.selected_editor ]; then
     mv $HOME/.selected_editor $HOME/.selected_editor~
+    ln -s $dotfiles/selected_editor $HOME/.selected_editor
 fi
-ln -s $dotfiles/selected_editor $HOME/.selected_editor
 
-if [ -f $HOME/.sudo_as_admin_successful ]; then
+if [ -f $HOME/.sudo_as_admin_successful ] && [ ! -L $HOME/.sudo_as_admin_successful ]; then
     mv $HOME/.sudo_as_admin_successful $HOME/.sudo_as_admin_successful~
+    ln -s $dotfiles/sudo_as_admin_successful $HOME/.sudo_as_admin_successful
 fi
-ln -s $dotfiles/sudo_as_admin_successful $HOME/.sudo_as_admin_successful
+
